+++ project/TestGradle/src/main/java/cl/principal/Application.java
@@ -6,9 +6,11 @@
 package cl.principal;
 
 import java.sql.SQLException;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.PropertySource;
 
 /**
  *
@@ -18,6 +20,8 @@ import org.springframework.context.annotation.ComponentScan;
 @SpringBootApplication
 public class Application {
     public static void main(String[] args) throws SQLException {
-            SpringApplication.run(Application.class, args);                
+            SpringApplication.run(Application.class, args);
+            //System.out.println("");
+            
     }
 }
