+++ project/TestGradle/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.driver-class-name = com.mysql.cj.jdbc.Driver
 spring.datasource.url = jdbc:mysql://localhost:3306/colegio
 spring.datasource.username = root
 spring.datasource.password = root
